The ECDSA Algorithm
Fixed the problem - the signature could not be verified for the message "message", since the signature was created for the message "signed message". The code should now run without errors and correctly verify the signature for the "signed message" message, as well as throw an error when trying to verify the signature for another message.